+++ lib/generator.dart
@@ -12,12 +12,11 @@ class SerializableGenerator extends GeneratorForAnnotation<Serializable> {
 
   @override
   Future<String> generateForAnnotatedElement(ClassElement element, Serializable annotation, BuildStep buildStep) async {
-
     var accessors = element.accessors;
     var getters = accessors.where((a) => a.kind == ElementKind.GETTER);
     var setters = accessors.where((a) => a.kind == ElementKind.SETTER);
 
-    return '''abstract class _\$${element.name}Serializable extends ISerializable {
+    return '''abstract class _\$${element.name}Serializable extends SerializableMap {
   ${getters.map((g) => 'get ${g.name}').join(';\n')};
   ${setters.map((g) => 'set ${g.displayName}(v)').join(';\n')};
 
@@ -25,14 +24,14 @@ class SerializableGenerator extends GeneratorForAnnotation<Serializable> {
     switch(key) {
       ${getters.map((a) => "case '${a.name}': return ${a.name};").join('\n')}
     }
-    throw new Exception('field not supported');
+    throwFieldNotFoundException(key, "${element.name}");
   }
 
   operator []=(String key, value) {
     switch(key) {
       ${setters.map((a) => "case '${a.displayName}': ${a.displayName} = value; return;").join('\n')}
     }
-    throw new Exception("The key \$key doesn't exist on class ${element.name}");
+    throwFieldNotFoundException(key, "${element.name}");
   }
 
   get keys => const [${element.fields.map((f) => "'${f.name}'").join(',')}];