+++ tool/build.dart
@@ -1,5 +1,5 @@
 import 'package:build_runner/build_runner.dart';
-import 'package:serializable/serializable_phase.dart';
+import 'package:serializable/phase.dart';
 
 
 main() async {