+++ CHANGELOG.md
@@ -1,7 +1,14 @@
-## 0.0.1
+## 0.1.0
 
-- Initial version
+- change name from `ISerializable` to `SerializableMap`
+- add `FieldNotFoundError`
+- rename `serializable_phase.dart` to `phase.dart`
+- rename `serializable_generator.dart` to `generator.dart`
 
 ## 0.0.2
 
-* add `toMap` adn `fromMap` methods to `ISerializable`
+- add `toMap` adn `fromMap` methods to `ISerializable`
+
+## 0.0.1
+
+- Initial version